+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -1,14 +1,13 @@
-package wooteco.subway.ui;
+package wooteco.subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
-import wooteco.subway.dto.StationRequest;
-import wooteco.subway.dto.StationResponse;
-import wooteco.subway.service.StationDto;
+import wooteco.subway.controller.dto.StationRequest;
+import wooteco.subway.controller.dto.StationResponse;
+import wooteco.subway.service.dto.StationDto;
 import wooteco.subway.service.StationService;
 
 import java.net.URI;
