+++ src/main/java/wooteco/subway/controller/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.controller.dto;
 
 public class StationRequest {
     private String name;
