+++ src/main/java/wooteco/subway/service/StationService.java
@@ -5,6 +5,7 @@ import java.util.stream.Collectors;
 
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
+import wooteco.subway.service.dto.StationDto;
 
 public class StationService {
