+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import wooteco.subway.dao.StationDao;
+import wooteco.subway.service.dto.StationDto;
 
 public class StationServiceTest {
